+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 
 * I wrote the code.
 
+* I added a UseEventAggregator method.
+
+* I added a PublishAsync method.
+